Enhancement: Add middleware to refresh session
