+++ middleware.ts
@@ -1,0 +1,10 @@
+import { createMiddlewareClient } from '@supabase/auth-helpers-nextjs'
+import type { NextRequest } from 'next/server'
+import { NextResponse } from 'next/server'
+
+export async function middleware(req: NextRequest) {
+  const res = NextResponse.next()
+  const supabase = createMiddlewareClient({ req, res })
+  await supabase.auth.getSession()
+  return res
+}
